refactor(ui): split basic tauri service domains
